refactor(cli): provider-registry bootstrap; drop Kalatori references
- Split the bootstrap monolith into commands/bootstrap/{types,shared,aws,gcp,
  aws-apps,index}.ts: a thin provider-registry dispatcher, shared helpers
  (exec/log/kind/waitForProviders/waitForCrds/synthAndApply — DRY), and one
  isolated module per provider. Adding a provider = implement BootstrapProvider
  and register it.
- Fix latent bug: the mgmt-cluster platform install now passes the mgmt
  kubeconfig to waitForProviders (was checking the Kind context).
- Remove Kalatori references from doc comments (infra/aws/cluster.ts, example/aws.ts).
