+++ packages/cli/src/commands/bootstrap/aws-apps.ts
